# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The JSON interface to a HyperUplink board, exposing the same\ncategories, forums, topics and administrative surfaces the web\ninterface serves, with every response shaped by the permissions\nof the user the API key belongs to.\n\nAuthentication happens with an API key that a signed-in user\nmints under Account -\u003e API on the board itself, and the secret\nit hands out once is sent either as a bearer token in the\nAuthorization header or verbatim in the X-API-Key header, so a\nrequest that carries neither is answered with 401 before it\never reaches a controller.",
        "title": "HyperUplink API",
        "contact": {
            "name": "HyperUplink",
            "url": "https://hyperup.link"
        },
        "license": {
            "name": "SEGV",
            "url": "https://xn--gckvb8fzb.com/segv/"
        },
        "version": "{{.Version}}"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/board/categories/{id}/movedown": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Move a category down",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The category identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
        "/admin/board/categories/{id}/moveup": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Move a category up",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The category identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/board/forums/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The forum identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The fields to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/forums.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "delete": {
                "description": "The delete is soft and propagates down to the topics and replies the\nforum holds.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a forum",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The forum identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
        "/admin/board/topics": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Show the topic settings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "topics": {
                                    "$ref": "#/definitions/setting.Topics"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update the topic settings",
                "parameters": [
                    {
                        "description": "The topic settings to store",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/topics.UpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/request.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/request.ValidationErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
        "/admin/comms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List the delivery sections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "sections": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "APIKeyAuth": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "forums.UpdateInput": {
            "type": "object",
            "required": [
                "category_id",
                "name",
                "slug"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 128,
                    "minLength": 0
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                },
                "slug": {
                    "type": "string",
                    "maxLength": 32,
                    "minLength": 1
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "forbidden"
                }
            }
        },
        "request.StatusResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "example": "ok"
                }
            }
        },
        "request.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
...
        "setting.Topics": {
            "type": "object",
            "properties": {
                "allow_kind_poll": {
                    "type": "boolean"
                },
                "allow_kind_question": {
                    "type": "boolean"
                },
                "allow_kind_rsvp": {
                    "type": "boolean"
                }
            }
        },
...
        "topics.UpdateInput": {
            "type": "object",
            "properties": {
                "allow_kind_poll": {
                    "type": "boolean"
                },
                "allow_kind_question": {
                    "type": "boolean"
                },
                "allow_kind_rsvp": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "description": "The API key secret on its own.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "The API key secret prefixed with \"Bearer \".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}